Restore full image optimization instructions in optimized URLs.
A very old previous commit (9741fa186) removed some of the optimization
configuration in the URL of optimized images. There was no explanation
given for this, and this has been linked to the cause of incorrect
optimization when processing a cache-miss through a CDN. This change
restores the previous url configuration as well as adds configuration for
the newer options such as small-screen and save-data.

Fixes #3681