Temporarily disable encoding of new fields for gradual deployment.
The new codes in the encoded URL will fail to be decoded on pagespeed
servers that don't have the updated code.  Thus, if an updated server
processes the HTML and rewrites urls with the new codes, when those urls
are randomly forwarded (by the CDN) to an old server, the image will fail
to render.  That's bad.

Instead, we'll deploy in two steps:
* First deploy the new url decoding code to all servers.
* Then deploy the new url encoding code to all servers.