Restore full image optimization instructions in optimized URLs. (#3)
* Restore full image optimization instructions in optimized URLs.

A very old previous commit (9741fa186) removed some of the optimization
configuration in the URL of optimized images. There was no explanation
given for this, and this has been linked to the cause of incorrect
optimization when processing a cache-miss through a CDN. This change
restores the previous url configuration as well as adds configuration for
the newer options such as small-screen and save-data.

Fixes #3681

* Add build script

* Temporarily disable encoding of new fields for gradual deployment.

The new codes in the encoded URL will fail to be decoded on pagespeed
servers that don't have the updated code.  Thus, if an updated server
processes the HTML and rewrites urls with the new codes, when those urls
are randomly forwarded (by the CDN) to an old server, the image will fail
to render.  That's bad.

Instead, we'll deploy in two steps:
* First deploy the new url decoding code to all servers.
* Then deploy the new url encoding code to all servers.