Fix #9 X-Mod-Pagespeed header ends with "null"
The X-Mod-Pagespeed header is populated partially via LASTCHANGE from src/build/lastchange.sh in the pagespeed build process. This script looks like they intended for local builders to modify it, but it is always fixed at zero. This fixes the earlier, failed attempt to put our own version number in that field.