Fix #9 X-Mod-Pagespeed header ends with "null"
The X-Mod-Pagespeed header is populated partially via LASTCHANGE from
src/build/lastchange.sh in the pagespeed build process. This script
looks like they intended for local builders to modify it, but it is
always fixed at zero. This fixes the earlier, failed attempt to put
our own version number in that field.

This is essentially equivalent to PR #10, but for the 1.12.34.2 branch
instead of 1.11.33.5.

Fixes #9 again.