Add fallback fetching of resources that are being optimized.
This adds a new pagespeed option "ModPagespeedFetcherFallbackTimeOutMs"

The problem causing 404s on websites with large CSS files was caused by the
call to BlockingFetch timing because the scheduler that was working on
optimizing the CSS file was taking too long.

In particular, this occurred when the CSS file referred to many other image
resources and the scheduler had those downloads queued up -- all subsequent
requests for the optimized file would initiate a refresh fetch to the underlying
resource, but that would get stuck in the scheduler queue behind all the other
resources to download.

Because of this, BlockingFetch would timeout after 5 seconds and return a 404
to the original client. That's bad.

Now, the code is modified so that after BlockingFetch starts the request for
the optimized resource it will start a second fallback request directly to the
unoptimized resource if the first request doesn't complete within a specified
timeout. This timeout is controlled by the ModPagespeedFetcherFallbackTimeOutMs
configuration setting.

The core logic in this change is encapsulated in the FetchRace class. This
class allows starting multiple background fetches in a race to write to the
output fetch. The first background fetch that starts writing a response wins.